requirements.txt
torch==1.8.0+cu101
torchvision==0.9.0+cu101
torchaudio==0.8.0
tensorboard==2.4.1
Pillow==7.1.0
openslide-python==1.1.2
numpy==1.17.4
tqdm

